# openapi.yaml
openapi: 3.0.3
info:
  title: SETEC eCommerce APIM
  version: 1.0.0
  description: REST API documentation for the SETEC eCommerce Customer App (Phase 1 MVP).
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server
  - url: https://staging-api.setec-ecommerce.com/api/v1
    description: Staging Server
  - url: https://api.setec-ecommerce.com/api/v1
    description: Production Server
paths:
...
  /users/me/reviews/:
    get:
      summary: List reviews created by current user
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20 }
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
...
  /products/{product_id}/reviews/:
    get:
      summary: List published reviews for product
      tags: [Products, Reviews]
      parameters:
        - in: path
          name: product_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20 }
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
    post:
      summary: Create review for product
      tags: [Products, Reviews]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: product_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5, example: 5 }
                title: { type: string, example: "Excellent Quality" }
                body: { type: string, example: "Loved the product, fits perfectly." }
                order_item_id: { type: string, format: uuid }
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        data:
          type: object
          description: Response payload
        meta:
          type: object
          properties:
            request_id: { type: string, nullable: true }
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
    ListEnvelope:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        meta:
          type: object
          properties:
            page: { type: integer, example: 1 }
            page_size: { type: integer, example: 20 }
            total: { type: integer, example: 50 }
            has_next: { type: boolean, example: true }
            request_id: { type: string, nullable: true }
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string, example: "Invalid input provided" }
        field: { type: string, nullable: true, example: "email" }